renaming app-greeting to app-diag: a simple diagnostic app to check selection, adapter
