+++ app/diag/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>app-greetings</name>
+	<name>app-diag</name>
 	<comment></comment>
 	<projects>
 	</projects>
