+++ app/diag/build.xml
@@ -1,8 +1,8 @@
-<project default="app-greetings">
+<project default="app-diag">
   <import file="../../dependencies/ant_settings.xml"/>
 
-  <target name="app-greetings" depends="compile-app">
-    <jar destfile="${build}/app-greetings-${version}.jar">
+  <target name="app-diag" depends="compile-app">
+    <jar destfile="${build}/app-diag-${version}.jar">
       <fileset dir="${classes}"/>
       <fileset dir="${src}" includes="**/*.fxml"/>
       <fileset dir="${resources}"/>
