+++ app/diag/pom.xml
@@ -1,11 +1,11 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <artifactId>app-greetings</artifactId>
+  <artifactId>app-diag</artifactId>
   <parent>
-  	<groupId>org.phoebus</groupId>
-  	<artifactId>app</artifactId>
-  	<version>0.0.1-SNAPSHOT</version>
+    <groupId>org.phoebus</groupId>
+    <artifactId>app</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
   </parent>
   <dependencies>
     <dependency>
