+++ app/diag/src/main/resources/META-INF/services/org.phoebus.ui.spi.MenuEntry
@@ -1,0 +1,1 @@
+org.phoebus.app.diag.DiagMenuEntry
