--- app/greetings/src/main/resources/META-INF/services/org.phoebus.framework.spi.AppDescriptor
@@ -1,1 +1,0 @@
-org.phoebus.applications.greetings.GreetingsApp
